# 8297476: Increase InlineSmallCode default from 1000 to 2500 for RISC-V