@RealFYang 2022-11-24
@shipilev @yhzhu20 : Thanks for the review!
I have also verified on both linux-x64 and linux-aarch64 platforms with following commands, still work as expected.

```
$ java -XX:+PrintFlagsFinal -version | grep InlineSmallCode
     intx InlineSmallCode                          = 2500                                {C2 pd product} {default}
openjdk version "20-internal" 2023-03-21
OpenJDK Runtime Environment (build 20-internal-adhoc.fyang.openjdk-jdk)
OpenJDK 64-Bit Server VM (build 20-internal-adhoc.fyang.openjdk-jdk, mixed mode, sharing)

$ java -XX:-TieredCompilation -XX:+PrintFlagsFinal -version | grep InlineSmallCode
     intx InlineSmallCode                          = 1000                                {C2 pd product} {default}
openjdk version "20-internal" 2023-03-21
OpenJDK Runtime Environment (build 20-internal-adhoc.fyang.openjdk-jdk)
OpenJDK 64-Bit Server VM (build 20-internal-adhoc.fyang.openjdk-jdk, mixed mode, sharing)
```

/integrate